QA-9309: Remove inconsistencies between encoded paths and names
The escaping of child names in getChildren(String) was relying on code
copied from JCRContentUtils.escapeLocalNodeName(String) which is doing
additional transformations of the input string (replacing some chars
with a whitespace), thus leading to inconsistencies between the name of
a node and the path to it. For VFS, we now rely on the same method to
escape illegal characters from a node path or a node name. The escaped
characters are the following ones: []*|:

Also moved escaping/unescaping methods to a dedicated utility class, and
added some unit tests.