QA-9309: Escape column characters from VFS paths and names (#13)
* Added column characters to the ones to be escaped from file paths and names. Replaced use of escaping methods from JCRContentUtils with new ones to restrict escaping of column character to the VFS provider.

* Remove inconsistencies between encoded paths and names
The escaping of child names in getChildren(String) was relying on code
copied from JCRContentUtils.escapeLocalNodeName(String) which is doing
additional transformations of the input string (replacing some chars
with a whitespace), thus leading to inconsistencies between the name of
a node and the path to it. For VFS, we now rely on the same method to
escape illegal characters from a node path or a node name. The escaped
characters are the following ones: []*|:

* Avoid having paths unescaped twice in VFS datasource